Code refactored - proper modules export
