+++ css_tasks/task11/js/elements.js
@@ -1,9 +1,4 @@
-// export const DOMElements = {
-//   service: document.querySelector('.service'),
-//   image: [...document.querySelectorAll('.service__element img')],
-// };
-
-export const resolver = new Promise((res) => {
+export const elements = new Promise((res) => {
   document.addEventListener('DOMContentLoaded', () => {
     const DOMElements = {
       service: document.querySelector('.service'),
